fix - 배포 관련 코드 삭제
