+++ src/constants/searchConstant.ts
@@ -1,0 +1,3 @@
+export const SEARCH_COOKIE = "searchs";
+
+export const SearchType = { region: "지역 검색", food: "음식 검색" } as const;
