+++ src/constants/typings/searchType.ts
@@ -1,0 +1,9 @@
+export interface KeywordType {
+  text: string;
+  id: number;
+}
+export interface RecordProps {
+  keywords: KeywordType[];
+  onRemoveKeyword: (id: number) => void;
+  onClearKeywords: () => void;
+}
