--- src/store/hooks/configstore.tsx
@@ -1,15 +1,0 @@
-import { createStore, combineReducers, applyMiddleware } from "redux";
-import thunk from "redux-thunk";
-import { createBrowserHistory } from "history";
-import userid from "../modules/userid";
-
-
-export const history = createBrowserHistory();
-
-const middlewares = [thunk];
-
-const enhancer = applyMiddleware(...middlewares);
-const rootReducer = combineReducers({ userid });
-const store = createStore(rootReducer, enhancer);
-
-export default store;
